requirements.txt
dearpygui==2.1.1
h5py==3.15.1
imgaug==0.4.0
ipython==8.12.3
keras==2.10.0
matplotlib==3.6.0
numpy==1.24.3
opencv-python==4.5.5.64
tensorflow==2.10.0
onnxruntime_gpu==1.23.2
Pillow==12.0.0
scipy
scikit-image
tensorflow==2.10.0
torch==2.9.0
